+++ lib/interfaces/common_row_properties.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-// Interface to ensure that all widgets implement this minimum
-// set of properties
-abstract class ICommonRowProperties extends Widget {
-  final bool showMaterialonIOS = null;
-  final bool visible = null;
-}